change name back to cdoFieldButton
